auth/api/urls.py
from django.urls import path
from rest_framework.authtoken.views import obtain_auth_token
from auth.api.views import registration_view, logout_view
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView


urlpatterns = [
    path("login/", view=obtain_auth_token, name="login"),
    path("register/", view=registration_view, name="register"),
    path("logout/", view=logout_view, name="logout"),
    path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
